Refactor main module into smaller components
